requirements.txt
asgiref==3.8.1
Django==5.1.4
django-leaflet==0.31.0
djangorestframework==3.15.2
leaflet==0.0.3
pillow==11.1.0
psycopg==3.2.3
python-dotenv==1.0.1
sqlparse==0.5.3
typing_extensions==4.12.2
gunicorn==20.1.0
dj-database-url==2.3.0
whitenoise==6.9.0